api/requirements.txt
fastapi
uvicorn
boto3
anthropic
pandas
pyarrow
python-dotenv
